yfinance-client/pyproject.toml
[tool.poetry]
name = "yfinance-client"
version = "0.1.0"
description = ""
authors = ["Danila <ddbihbka@gmail.com>"]
readme = "README.md"
packages = [{include = "yfinance_client"}]

[tool.poetry.dependencies]
python = "<3.13,>=3.11"
yfinance = "^0.2.28"
scipy = "^1.11.1"
stock-monitor-data = {path = "../stock-monitor-data"}
kafka-python = "^2.0.2"
rich = "^13.5.2"

[tool.poetry.scripts]
yfinance_client = "yfinance_client.main:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
